feat: Phase 3 - 服务层拆分 (HermesService → 7个专职服务 + Facade)
将 HermesService (1,689行/66方法) 拆分为 7 个专职服务 + Facade 门面:

新增服务:
- SessionService (535行): 会话CRUD+搜索+导出+FTS5
- SkillService (311行): 技能CRUD+元数据解析+市场
- MemoryService (443行): 记忆读写+画像+学习记录+知识提取
- CronService (146行): 定时任务CRUD+调度
- AnalyticsService (333行): 统计+趋势+系统状态+配置
- AgentService (88行): 子Agent管理
- ToolService (172行): 工具列表+MCP网关

改造:
- HermesService: 1,689行 → 264行 (-84%)
- 所有66个方法签名保持不变 (向后兼容)
- 路由层/前端零修改

测试: 7个服务导入✓ Facade方法完整✓ 调用正常✓ 96工具发现✓

关联 Issue: #22 #23 #24 #25
